Добавил тесты к новым компонентам и обновил старые
